+++ src/test/java/ladder/domain/ladder/LadderLineTest.java
@@ -1,6 +1,5 @@
-package ladder.domain;
+package ladder.domain.ladder;
 
-import ladder.domain.Ladder.LadderLine;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.*;
